// src/types/quotationMargin.ts
export type OptionsFilterQuotationMargin = {
   region: any[];
   subRegion: any[];
   delivery: any[];
   type: any[];
   orderNumber: any[];
   modelCode: any[];
   series: any[];
};

export type OptionFilterFromCalculateFile = {
   type: any[];
   orderNumber: any[];
   modelCode: any[];
   series: any[];
};